[clean] Separete configurations into two files
